todo ok landing gian instalado
