+++ src/components/Landing/Landing.scss
@@ -59,6 +59,7 @@
     border-bottom-right-radius: 28px;
     background-color: transparent;
     margin: 25px;
+    text-decoration: none;
 }
